11/05 요청페이지&요청(Submit)

+++ src/main/java/com/springproject/dtos/MeasureDescriptionDto.java
@@ -1,11 +1,11 @@
 package com.springproject.dtos;
 
 public class MeasureDescriptionDto {
-	private int acceptNo;
-	public int getAcceptNo() {
+	private Long acceptNo;
+	public Long getAcceptNo() {
 		return acceptNo;
 	}
-	public void setAcceptNo(int acceptNo) {
+	public void setAcceptNo(Long acceptNo) {
 		this.acceptNo = acceptNo;
 	}
 	public String getMeasureDescription() {
